extract line scaling into a function
